+++ common/src/main/java/whocraft/tardis_refined/compat/create/TardisDisplaySource.java
@@ -12,6 +12,7 @@ import whocraft.tardis_refined.common.blockentity.device.FlightDetectorBlockEnti
 import whocraft.tardis_refined.common.capability.tardis.TardisLevelOperator;
 import whocraft.tardis_refined.common.tardis.TardisNavLocation;
 import whocraft.tardis_refined.common.util.MiscHelper;
+import whocraft.tardis_refined.common.util.Platform;
 import whocraft.tardis_refined.compat.ModCompatChecker;
 import whocraft.tardis_refined.compat.valkyrienskies.VSHelper;
 import whocraft.tardis_refined.constants.ModMessages;
@@ -37,7 +38,7 @@ public class TardisDisplaySource extends DisplaySource {
                 if (isPresent) {
                     TardisLevelOperator levelOperator = TardisLevelOperator.get(serverLevel).get();
                     TardisNavLocation currentLoc = levelOperator.getPilotingManager().getCurrentLocation();
-                    if (ModCompatChecker.valkyrienSkies()) {
+                    if (ModCompatChecker.valkyrienSkies() && Platform.getServer() != null) {
                         currentLoc = VSHelper.toWorldLocation(currentLoc);
                     }