+++ changelog.md
@@ -43,3 +43,4 @@
 - Bug fix: Players can take damage or die while in shell view.
 - Bug fix: Players can receive new potion effects while in shell view.
 - Bug fix: Create TARDIS Summary does not display all lines correctly.
+- Bug fix: Crash when closing the game near a TARDIS Create display while Valkyrien Skies is installed.